Added parameter distribution container
