+++ setup.py
@@ -24,6 +24,7 @@ setup(
     entry_points={
         'webviz_config_containers': [
             'ParameterDistribution = webviz_subsurface.containers:ParameterDistribution',
+            'ParameterCorrelation = webviz_subsurface.containers:ParameterCorrelation',
             'DiskUsage = webviz_subsurface.containers:DiskUsage',
             'SubsurfaceMap = webviz_subsurface.containers:SubsurfaceMap',
             'HistoryMatch = webviz_subsurface.containers:HistoryMatch',
